Rework Proxy & Route interfaces
- Proxy now only exposes hostnames nad creates routes on demand only
- Route now has only one hostname and is responsible for httpx client creation
