+++ testsuite/openshift/envoy.py
@@ -1,14 +1,12 @@
 """Module containing all Envoy Classes"""
-from functools import cached_property
 from importlib import resources
 
 from openshift import Selector
 
-from testsuite.httpx import HttpxBackoffClient
 from testsuite.openshift.client import OpenShiftClient
 from testsuite.openshift.httpbin import Httpbin
 from testsuite.openshift.objects.proxy import Proxy
-from testsuite.openshift.objects.route import OpenshiftRoute, Route
+from testsuite.openshift.objects.route import OpenshiftRoute
 
 
 # pylint: disable=too-many-instance-attributes
@@ -28,31 +26,15 @@ class Envoy(Proxy):
 
         self.envoy_objects: Selector = None  # type: ignore
 
-    @cached_property
-    def route(self) -> Route:
-        """Returns route for object"""
-        with self.openshift.context:
-            return (
-                self.envoy_objects.narrow("route")
-                .narrow(lambda route: route.model.metadata.name == self.name)
-                .object(cls=OpenshiftRoute)
-            )
-
-    def add_hostname(self, name) -> str:
+    def expose_hostname(self, name) -> OpenshiftRoute:
         """Add another hostname that points to this Envoy"""
-        route = OpenshiftRoute(dict_to_model=self.openshift.routes.expose(name, self.name).as_dict())
+        route = OpenshiftRoute.create_instance(
+            self.openshift, name, service_name=self.name, target_port="web", labels={"app": self.label}
+        )
+        route.commit()
         with self.openshift.context:
             self.envoy_objects = self.envoy_objects.union(route.self_selector())
-        return route.hostnames[0]
-
-    @cached_property
-    def hostname(self):
-        """Returns hostname of this envoy"""
-        return self.route.hostnames[0]
-
-    def client(self, **kwargs):
-        """Return Httpx client for the requests to this backend"""
-        return HttpxBackoffClient(base_url=f"http://{self.hostname}", **kwargs)
+        return route
 
     def commit(self):
         """Deploy all required objects into OpenShift"""
@@ -97,9 +79,21 @@ class TLSEnvoy(Envoy):
         self.backend_ca_secret = envoy_ca_secret
         self.envoy_cert_secret = envoy_cert_secret
 
-    def client(self, **kwargs):
-        """Return Httpx client for the requests to this backend"""
-        return HttpxBackoffClient(base_url=f"https://{self.hostname}", **kwargs)
+    def expose_hostname(self, name) -> OpenshiftRoute:
+        """Add another hostname that points to this Envoy"""
+        route = OpenshiftRoute.create_instance(
+            self.openshift,
+            name,
+            service_name=self.name,
+            target_port="web",
+            labels={"app": self.label},
+            tls=True,
+            termination="passthrough",
+        )
+        route.commit()
+        with self.openshift.context:
+            self.envoy_objects = self.envoy_objects.union(route.self_selector())
+        return route
 
     def commit(self):
         self.envoy_objects = self.openshift.new_app(
