+++ testsuite/openshift/objects/route.py
@@ -2,6 +2,9 @@
 from abc import ABC, abstractmethod
 from functools import cached_property
 
+from httpx import Client
+
+from testsuite.httpx import HttpxBackoffClient
 from testsuite.openshift.objects import OpenShiftObject
 
 
@@ -10,13 +13,48 @@ class Route(ABC):
 
     @cached_property
     @abstractmethod
-    def hostnames(self) -> list[str]:
-        """Returns all Route hostnames"""
+    def hostname(self) -> str:
+        """Returns one of the Route valid hostnames"""
+
+    @abstractmethod
+    def client(self, **kwargs) -> Client:
+        """Return Httpx client for the requests to this backend"""
 
 
 class OpenshiftRoute(OpenShiftObject, Route):
     """Openshift Route object"""
 
+    @classmethod
+    def create_instance(
+        cls,
+        openshift,
+        name,
+        service_name,
+        target_port: int | str,
+        tls=False,
+        termination="edge",
+        labels: dict[str, str] = None,
+    ):
+        """Creates new OpenshiftRoute instance"""
+        model = {
+            "apiVersion": "route.openshift.io/v1",
+            "kind": "Route",
+            "metadata": {"name": name, "labels": labels},
+            "spec": {
+                "to": {"name": service_name, "kind": "Service"},
+                "port": {"targetPort": target_port},
+            },
+        }
+        if tls:
+            model["spec"]["tls"] = {"termination": termination}  # type: ignore
+        return cls(model, context=openshift.context)
+
+    def client(self, **kwargs) -> Client:
+        protocol = "http"
+        if "tls" in self.model.spec:
+            protocol = "https"
+        return HttpxBackoffClient(base_url=f"{protocol}://{self.hostname}", **kwargs)
+
     @cached_property
-    def hostnames(self):
-        return [self.model.spec.host]
+    def hostname(self):
+        return self.model.spec.host
